:cake: Update
* Refactored code
* Added event system
* Added editorconfig
* Moved source code
* Updated index.html
* Updated package
* Fixed minification
