+++ .editorconfig
@@ -1,0 +1,4 @@
+root = true
+indent_style = space
+indent_size = 2
+end_of_line = lf
