+++ .eslintrc
@@ -9,6 +9,7 @@
     "no-param-reassign": 0,
     "default-case": 0,
     "consistent-return": 0,
-    "arrow-body-style": 0
+    "arrow-body-style": 0,
+    "linebreak-style": 0
   }
 }
