+++ gulpfile.js
@@ -1,20 +1,19 @@
 const gulp = require('gulp');
-const babel = require('gulp-babel');
-const minify = require('gulp-babel-minify');
+const uglify = require('gulp-uglify');
 const rename = require('gulp-rename');
-
+const browserify = require('browserify');
+const source = require('vinyl-source-stream');
 gulp.task('babel', () => {
-  return gulp.src('index.js')
-  .pipe(babel({
-    presets: ['es2015'],
-  }))
-  .pipe(rename('smallfoot.js'))
+  return browserify({ entries: './index.js' })
+  .transform('babelify', { presets: ['es2015'] })
+  .bundle()
+  .pipe(source('smallfoot.js'))
   .pipe(gulp.dest('dist/'));
 });
 
 gulp.task('minify', () => {
-  return gulp.src('index.js')
-    .pipe(minify())
+  return gulp.src('dist/smallfoot.js')
+    .pipe(uglify())
     .pipe(rename('smallfoot.min.js'))
     .pipe(gulp.dest('dist/'));
 });
