+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "smallfoot",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "description": "A small and cleaner version of bigfoot.js",
   "main": "index.js",
   "scripts": {
@@ -24,6 +24,7 @@
     "babel-eslint": "^6.1.2",
     "babel-minify": "^0.1.12",
     "babel-preset-es2015": "^6.13.2",
+    "browserify": "^13.1.0",
     "eslint": "^3.2.2",
     "eslint-config-airbnb": "^10.0.0",
     "eslint-plugin-import": "^1.12.0",
@@ -33,6 +34,8 @@
     "gulp-babel": "^6.1.2",
     "gulp-babel-minify": "^0.1.12",
     "gulp-minify": "0.0.14",
-    "gulp-rename": "^1.2.2"
+    "gulp-rename": "^1.2.2",
+    "gulp-uglify": "^2.0.0",
+    "vinyl-source-stream": "^1.1.0"
   }
 }
